Create Find the number of unique combinations of beads
This is an exercise for a coding test recently undertaken by myself. The initial problem was as follows:

A child has a bunch of beads that they want to make a necklace out of, these beads are labeled  by a positive integer. 
The child wants each bead to add together with the proceeding bead to add up to a prime number. 
The child wishes to know how many unique necklace iterations can be made, with the addition of the end integer combining with the first bead of the next list of beads.

ex. 
Unique list - 1,2,3,4 + 1,2,4,3